Rename prefix to clf_
rename file prefix from ra_ to clf_ to better reflect the mission of this
project.
