+++ cczoo/cross_lang_framework/clf_client/app/java.manifest.template
