+++ cczoo/cross_lang_framework/clf_client/app/test.manifest.template
