+++ cczoo/cross_lang_framework/clf_client/java/1/gramine_jni.manifest.template
