+++ cczoo/cross_lang_framework/clf_server/Makefile
@@ -18,7 +18,7 @@ CFLAGS += -fPIE
 LDFLAGS += -pie
 
 .PHONY: all
-all: ra_server
+all: clf_server
 
 
 CFLAGS += -Wall -std=c11 -I$(GRAMINEDIR)/Pal/src/host/Linux-SGX/tools/ra-tls \
@@ -29,12 +29,12 @@ LDFLAGS += $(shell pkg-config --libs mbedtls_gramine)
 # linker option --no-as-needed is required because SGX DCAP library (libsgx_dcap_quoteverify.so)
 # does dlopen() instead of directly linking against libsgx_urts.so, and without this option
 # compilers remove the "seemingly unused" libsgx_urts.so
-ra_server: ra_server.c interact.c fop.c
+clf_server: clf_server.c interact.c fop.c
 	$(CC) $^ $(CFLAGS) $(LDFLAGS) -Wl,--no-as-needed -lsgx_urts -lsecret_prov_verify_dcap -pthread -o $@
 
 
 .PHONY: clean
 clean:
 	$(RM) -r \
-		ra_server
+		clf_server
