+++ cczoo/cross_lang_framework/clf_server/README.md
@@ -4,7 +4,7 @@
 
 RA_TLS_ALLOW_DEBUG_ENCLAVE_INSECURE=1 \
 RA_TLS_ALLOW_OUTDATED_TCB_INSECURE=1 \
-./ra_server &
+./clf_server &
 
 kill %%
 ```
