--- cczoo/cross_lang_framework/ra_server/files/README
@@ -1,7 +1,0 @@
-This directory contains auto-generated Protected Files, encrypted via pf_crypt.
-In particular, it contains:
-
-- `plain.txt` -- file to encrypt
-- `wrap-key` -- pre-generated wrap key
-
-Also, `input.txt` is generated by encoding `plain.txt` using `wrap-key`.
