--- cczoo/cross_lang_framework/ra_server/files/plain.txt
@@ -1,1 +1,0 @@
-helloworld
